replace client.go with new-client.go
